refactor: continue eliminating numpy-to-python conversions with native backend operations
This commit continues the migration from numpy-based array conversions to native backend operations across multiple modules:

- Add arctan and lgamma functions to all backends (cuda, jax, mlx) and port definitions
- Replace numpy array conversions with backend.to_scalar and backend array operations in entropy, geometry, safety, and thermodynamics modules
- Refactor manifold curvature tensor computations to use vectorized backend operations instead of Python loops
- Replace Python math module functions with backend-native equivalents (lgamma_scalar, acos_scalar, pi_value, e_value, sqrt_scalar, log_scalar, exp_scalar)
- Update weight serialization in adapter service and merge helpers to use native MLX/JAX safetensors serialization
- Add struct-based hashing in computation cache for improved performance
- Keep arrays on GPU throughout computation pipeline instead of converting to numpy/python

This improves performance by avoiding expensive device-to-host transfers and enables proper operation with non-numpy backends like MLX.

+++ src/modelcypher/core/use_cases/merge/stages/probe.py
@@ -896,7 +896,7 @@ def _extract_top_k_dims(
     # Convert to Python using backend to_scalar - no NumPy
     n_top = int(top_indices_arr.shape[0])
     top_indices = sorted([int(b.to_scalar(top_indices_arr[i])) for i in range(n_top)])
-    indices_arr = b.array(top_indices)
+    indices_arr = b.array(top_indices, dtype="int32")
     selected_acts = b.take(activation_vector, indices_arr, axis=0)
     selected_abs = b.take(abs_vals, indices_arr, axis=0)
     b.eval(selected_acts, selected_abs)
